Move unit test support modules under ./lib/test
