+++ lib/plugins/cast-spec.js
@@ -2,7 +2,7 @@
 'use strict'
 
 const _ = require('lodash')
-const { Logbus } = require('../../test/unit/logbus')
+const { Logbus } = require('../test/logbus')
 const Plugin = require('./cast')
 
 describe('plugin-cast', () => {
