+++ lib/plugins/drop-fields-spec.js
@@ -1,7 +1,7 @@
 
 'use strict'
 
-const { Logbus } = require('../../test/unit/logbus')
+const { Logbus } = require('../test/logbus')
 const Plugin = require('./drop-fields')
 
 describe('plugin-drop-fields', () => {
