+++ lib/plugins/input/elasticsearch-spec.js
@@ -2,7 +2,7 @@
 'use strict'
 
 const _ = require('lodash')
-const { Logbus } = require('../../../test/unit/logbus')
+const { Logbus } = require('../../test/logbus')
 const Plugin = require('./elasticsearch')
 const nock = require('nock')
