+++ lib/plugins/keep-fields-spec.js
@@ -1,7 +1,7 @@
 
 'use strict'
 
-const { Logbus } = require('../../test/unit/logbus')
+const { Logbus } = require('../test/logbus')
 const Plugin = require('./keep-fields')
 
 describe('plugin-keep-fields', () => {
