+++ lib/test/logbus.js
@@ -41,7 +41,7 @@ const Logbus = (stage) => {
   })
 
   // pretend entire pipeline started
-  ready = true
+  let ready = true
 
   // helper for test cases
   async function wait(ms, f) {
